SPL: use generic mips start.S file for SPL
This will call the SPL spcific board_init_f function
implemented in the renamed spl_pistachio.c.
While here I've cleaned up the unnecessary build of
mips32 (no longer used) and the division of tasks
for the SPL.
