+++ arch/mips/cpu/Makefile
@@ -2,10 +2,9 @@
 # SPDX-License-Identifier:	GPL-2.0+
 #
 
-ifneq ($(CONFIG_SPL_BUILD),y)
 extra-y	:= start.o
-endif
 
 obj-y	+= cpu.o
 obj-y	+= interrupts.o
+obj-y	+= mmu.o
 obj-y	+= time.o
