--- arch/mips/cpu/mips32/Makefile
@@ -1,9 +1,0 @@
-#
-# SPDX-License-Identifier:	GPL-2.0+
-#
-
-ifeq ($(CONFIG_SPL_BUILD),y)
-extra-y			:= start.o
-endif
-
-obj-y			+= mmu.o
