--- board/imgtec/pistachio_bub/lowlevel_init.h
@@ -1,15 +1,0 @@
-/*
- * Low level initialization code header file for MIPS32 CPU-core SPL
- *
- * Copyright (c) 2015 Imagination Technologies
- *
- * SPDX-License-Identifier:	GPL-2.0+
- */
-
-#ifndef __IMGTEC_PISTACHIO_LOWLEVEL_INIT_H_
-#define __IMGTEC_PISTACHIO__LOWLEVEL_INIT_H_
-
-void spl_lowlevel_init(void);
-void spl_end(void);
-
-#endif
